refactor(web): read plugin profile from query cache
