+++ web/features/agent-v2/agent-detail/configure/components/preview/__tests__/versions-panel.spec.tsx
@@ -1,6 +1,6 @@
 import type { AgentConfigSnapshotSummaryResponse } from '@dify/contracts/api/console/agent/types.gen'
 import { fireEvent, screen } from '@testing-library/react'
-import { render } from '@/test/console/render'
+import { renderWithAccountProfile as render } from '@/test/console/account-profile'
 import { AgentPreviewVersionsPanel } from '../versions-panel'
 
 const versions: AgentConfigSnapshotSummaryResponse[] = [
@@ -47,6 +47,13 @@ vi.mock('@/hooks/use-timestamp', () => ({
 
 vi.mock('@/service/client', () => ({
   consoleQuery: {
+    account: {
+      profile: {
+        get: {
+          queryKey: () => [['console', 'account', 'profile', 'get'], { type: 'query' }],
+        },
+      },
+    },
     agent: {
       byAgentId: {
         versions: {
@@ -58,17 +65,6 @@ vi.mock('@/service/client', () => ({
     },
   },
 }))
-
-vi.mock('@/context/account-state', async () => {
-  const { createAccountStateModuleMock } = await import('@/test/console/state-fixture')
-  return createAccountStateModuleMock(() => ({
-    userProfile: {
-      id: 'user-1',
-      name: 'Alice',
-      email: 'alice@example.com',
-    },
-  }))
-})
 
 describe('AgentPreviewVersionsPanel', () => {
   beforeEach(() => {
@@ -86,6 +82,7 @@ describe('AgentPreviewVersionsPanel', () => {
           onSelectVersion={handleSelectVersion}
           onClose={vi.fn()}
         />,
+        { accountProfile: { id: 'user-1', name: 'Alice', email: 'alice@example.com' } },
       )
 
       fireEvent.click(screen.getByRole('button', { name: /Initial release/i }))
@@ -103,6 +100,7 @@ describe('AgentPreviewVersionsPanel', () => {
           onSelectVersion={handleSelectVersion}
           onClose={vi.fn()}
         />,
+        { accountProfile: { id: 'user-1', name: 'Alice', email: 'alice@example.com' } },
       )
 
       fireEvent.click(screen.getByRole('button', { name: /currentDraft/i }))
@@ -120,6 +118,7 @@ describe('AgentPreviewVersionsPanel', () => {
           onSelectVersion={vi.fn()}
           onClose={vi.fn()}
         />,
+        { accountProfile: { id: 'user-1', name: 'Alice', email: 'alice@example.com' } },
       )
 
       fireEvent.click(screen.getByRole('button', { name: /filter/i }))
@@ -137,6 +136,7 @@ describe('AgentPreviewVersionsPanel', () => {
           onSelectVersion={vi.fn()}
           onClose={vi.fn()}
         />,
+        { accountProfile: { id: 'user-1', name: 'Alice', email: 'alice@example.com' } },
       )
 
       fireEvent.click(screen.getByRole('button', { name: /filter/i }))
